# Optimize rand.Intn(65536) to rand.Int63() >> 47,  before:22.6ns/op, after:18.6ns/op