--- a/common/dice/dice.go
+++ b/common/dice/dice.go
@@ -25,7 +25,7 @@ func RollDeterministic(n int, seed int64) int {
 
 // RollUint16 returns a random uint16 value.
 func RollUint16() uint16 {
-	return uint16(rand.Int31() >> 15)
+	return uint16(rand.Int63() >> 47)
 }
 
 func RollUint64() uint64 {