Heavily inspired by math/rand. with ~20% performance improvement (BenchmarkInt31 vs. BenchmarkIntn).

\# go test -bench=. -benchmem -run=none
goos: darwin
goarch: amd64
pkg: v2ray.com/core/common/dice
BenchmarkRoll1-4    	1000000000	         0.329 ns/op	       0 B/op	       0 allocs/op
BenchmarkRoll20-4   	43471560	        24.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkIntn1-4    	52293325	        22.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkIntn20-4   	46995854	        24.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkInt31-4    	62762128	        18.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkIntn-4     	52576292	        22.6 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	v2ray.com/core/common/dice	9.313s