@rprx 2020-11-14
看起来 rand.Int31() 的实现是这样的：

```
// Int31 returns a non-negative pseudo-random 31-bit integer as an int32.
func (r *Rand) Int31() int32 { return int32(r.Int63() >> 32) }
```

如果 32 位机器上也是这样的话，试试一步到位 Int63()？