@chinsyo 2020-11-14
> 看起来 rand.Int31() 的实现是这样的：
> 
> ```
> // Int31 returns a non-negative pseudo-random 31-bit integer as an int32.
> func (r *Rand) Int31() int32 { return int32(r.Int63() >> 32) }
> ```
> 
> 如果 32 位机器上也是这样的话，试试一步到位 rand.Int63()？

在我的 mbpr early 2015 进行性能测试，除首次Int63有明显性能优势外，表现非常接近。

首次：
\#go test -bench=. -benchmem -run=none
goos: darwin
goarch: amd64
pkg: v2ray.com/core/common/dice
BenchmarkRoll1-4    	1000000000	         0.326 ns/op	       0 B/op	       0 allocs/op
BenchmarkRoll20-4   	48462705	        24.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkIntn1-4    	45738037	        25.7 ns/op	       0 B/op	       0 allocs/op
BenchmarkIntn20-4   	45634132	        25.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkInt31-4    	50661666	        22.2 ns/op	       0 B/op	       0 allocs/op
BenchmarkInt63-4    	56532728	        18.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkIntn-4     	46046926	        25.4 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	v2ray.com/core/common/dice	7.888s

其余：
\#go test -bench=. -benchmem -run=none
goos: darwin
goarch: amd64
pkg: v2ray.com/core/common/dice
BenchmarkRoll1-4    	1000000000	         0.332 ns/op	       0 B/op	       0 allocs/op
BenchmarkRoll20-4   	46905194	        25.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkIntn1-4    	50883860	        22.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkIntn20-4   	43468287	        24.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkInt31-4    	60842923	        18.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkInt63-4    	63068659	        18.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkIntn-4     	52297114	        22.1 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	v2ray.com/core/common/dice	7.492s

相关内容已提交，请重新review。