@chinsyo 2020-11-14
> 现在只剩一个小问题：
> 
> common/dice/dice.go 的 init 已调用 rand.Seed，dice_test.go 新增的函数应该无需重新调用它（比如 dice_test.go 原有的函数）

Fixed, it looks like performace is getting better.

\#go test -bench=. -benchmem -run=none
goos: darwin
goarch: amd64
pkg: v2ray.com/core/common/dice
BenchmarkRoll1-4    	1000000000	         0.339 ns/op	       0 B/op	       0 allocs/op
BenchmarkRoll20-4   	44506482	        24.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkIntn1-4    	45430833	        25.4 ns/op	       0 B/op	       0 allocs/op
BenchmarkIntn20-4   	45414666	        24.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkInt31-4    	52839327	        21.9 ns/op	       0 B/op	       0 allocs/op
BenchmarkInt63-4    	62825142	        18.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkIntn-4     	44667235	        26.0 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	v2ray.com/core/common/dice	7.896s